ordering: u64 priority + select_nth_unstable cap (R-05)
Extend priority key from u32 to u64 by packing the move's (q, r) into
the low 32 bits (i16-bitcast-to-u16 halves). Two distinct legal Coords
have distinct keys, so the key is total — no ties at the sort site.

Replace the full stable sort + take(MOVE_GEN_CAP) pattern with
select_nth_unstable_by_key (O(n) partition) followed by sort_by_key
over the surviving prefix. Total-order key makes the top-N selection
deterministic without requiring sort stability.

Widen SearchScratch.scored tuple from (u32, u8, Coord) to
(u64, u8, Coord) accordingly.

Rebaseline event: tie order at the ordering sort site is now decided
by Coord pack, not generation order, so per-fixture node counts drift.
Strength is unchanged — search correctness is unaffected and game-tree
values per move are identical. The forced_block_of_opp_closed_five
test asserted a single root move; the engine now legitimately defers
the block to stone 2 of the same turn and develops on stone 1 (higher
eval, same final outcome). Test relaxed to assert the non-losing
score envelope instead of pinning a specific root move.

See SPEC_ENGINE.md "Ordering — encoding" and the superseded Phase 16
node-count note for the rebaseline rationale.

+++ hammerhead-engine/src/search.rs
@@ -49,7 +49,7 @@ pub struct SearchScratch {
     /// early-return path, where it never touches its own slot.
     pub moves: Box<[Vec<Coord>; MAX_PLY]>,
     /// Per-ply sort scratch for `order_moves_with_buckets`.
-    pub scored: Box<[Vec<(u32, u8, Coord)>; MAX_PLY]>,
+    pub scored: Box<[Vec<(u64, u8, Coord)>; MAX_PLY]>,
     /// Per-ply bucket-value output array (parallel to `moves` after
     /// ordering).
     pub buckets: Box<[Vec<u8>; MAX_PLY]>,
